+++ src/openfilings/cli.py
@@ -47,6 +47,7 @@ class SourceOption(StrEnum):
     sedar = "sedar"
     smv = "smv"
     sfc = "sfc"
+    asx = "asx"
 
 
 class OcrOption(StrEnum):